Provide a tool to read blog entries via the WordPress REST API.
The `bwfeeds` utility is now superfluous.

The index format has been updated to support multiple categories per post.

This is superior to reading RSS streams because the data returned by the
API are intended to be consumed by a machine and pagination works reliably.
We missed some posts when using the RSS feed due to the limitation on the
number of items per feed.